requirements.txt

pandas
opencv-python-headless==4.8.1.78
tensorflow==2.15
numpy
Pillow
protobuf
fastapi
uvicorn
aiortc
streamlit-webrtc



